dom/base/test/chrome/test_bug339494.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=339494
-->
<window title="Mozilla Bug 339494"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
  xmlns:xul="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>      

<body  xmlns="http://www.w3.org/1999/xhtml">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=339494">Mozilla Bug 339494</a>
<p id="display"></p>
<div id="content" style="display: none">
  <xul:hbox id="d"/>
  <xul:hbox id="s"/>
</div>
<pre id="test">
</pre>
</body>

<script class="testbody" type="application/javascript">

/** Test for Bug 339494 **/

  var d = document.getElementById("d");
  
  d.setAttribute("hhh", "testvalue");

  document.addEventListener("DOMAttrModified", removeItAgain, false);
  d.removeAttribute("hhh");
  document.removeEventListener("DOMAttrModified", removeItAgain, false);

  function removeItAgain()
  {
    ok(!d.hasAttribute("hhh"), "Value check 1",
       "There should be no value");
    isnot(d.getAttribute("hhh"), "testvalue", "Value check 2");
    document.removeEventListener("DOMAttrModified", removeItAgain, false);
    d.removeAttribute("hhh");
    ok(true, "Reachability", "We shouldn't have crashed");
  }
  
  var s = document.getElementById("s");
  
  s.setAttribute("ggg", "testvalue");

  document.addEventListener("DOMAttrModified", compareVal, false);
  s.setAttribute("ggg", "othervalue");
  document.removeEventListener("DOMAttrModified", compareVal, false);

  function compareVal()
  {
    ok(s.hasAttribute("ggg"), "Value check 3",
       "There should be a value");
    isnot(s.getAttribute("ggg"), "testvalue", "Value check 4");
    is(s.getAttribute("ggg"), "othervalue", "Value check 5");
  }

</script>

</window>
